refactor(addLocation) : transformation de AddLocation en LocationForm, et séparation des responsabilités
